Replace regexp matching on URL with path.split.
The original regexp had several issues including not working with the
config.URL and matching 'index.html' on 'index.htmlnotreally'.

Also remove the early 404 and full Swagger UI file list. This is flakey,
liable to change and error-prone, and entirely unnecessary as this is
already handled by the swaggerFiles package.